repositories/latent-diffusion/requirements.txt
Pillow
torch
pytorch_lightning
torchvision
einops
numpy
omegaconf